Added unused attribute to many unused function parameters

+++ src/include/utils.h
@@ -678,8 +678,8 @@ STRENGTH_APPLY_INDEX(struct creature *ch)
 #define GET_OBJ_COST(obj)        ((obj)->shared->cost)
 // can only be used to get the rent, not set
 #define GET_OBJ_RENT(obj)        ((obj)->plrtext_len ? \
-                                 ((obj)->plrtext_len << 3) : \
-                                 (obj)->shared->cost_per_day)
+                                  ((int)(obj)->plrtext_len << 3) :  \
+                                  (obj)->shared->cost_per_day)
 
 #define GET_OBJ_EXTRA(obj)        ((obj)->obj_flags.extra_flags)
 #define GET_OBJ_EXTRA2(obj)        ((obj)->obj_flags.extra2_flags)
